+++ apps/mobile-wallet/src/api/queries/dAppQueries.ts
@@ -1,0 +1,15 @@
+import { queryOptions } from '@tanstack/react-query'
+import axios from 'axios'
+
+import { DApp } from '~/features/ecosystem/ecosystemTypes'
+
+interface DAppsQueryOptions<T> {
+  select: (dApps: DApp[]) => T
+}
+
+export const dAppQueries = <T>({ select }: DAppsQueryOptions<T>) =>
+  queryOptions({
+    queryKey: ['dApps'],
+    queryFn: () => axios.get('https://publicapi.alph.land/api/dapps').then((res) => res.data),
+    select
+  })
